fix(filtering): avoid segfault in body filtering
In some cases apache2 drop the request before the body has finish to filter, in those case our context cleanup
would drop the filter while it's used resulting in a segfault.

Also we create the body filter as soon as possible so it will be available even if the request is drop from apache.

Made also some ajustements in case the filtering does not work to correctly free up created buffer, this should result in less memory leaks
(or event not) in case of errors during the filtering.